# file_scan.c: avoid work when processing inline-only small files